[Fix][Refactor] Fixed ConcurrentModificationException on SseKeepAliveEvent and IOException on app termination
A ConcurrentModificationException was being called when the a node leaves (refresh or app termination) which causes the node to be removed from the clusters, while the keep-alive event was being sent, which iterates over the cluster every once in a while, so when a node is removed, the collection which is being iterated over is changed. It was fixed by using an iterator instead of an enhanced for loop, and using iterator.remove(). An IOException was getting thrown when a node leaves which is caused by the spontaneous exit while the SSE HTTP connection is open. This was fixed by closing the connection on the client-end whenever a user refreshes or terminates the application.